breaking the admin page to components
